chore(source-jira): Merge timezone test and consolidate into test_issues.py
- Merged timezone test from old test_issues.py into test_issues_stream.py
- Renamed test_issues_stream.py to test_issues.py
- Deleted old test_issues.py
- All 50 tests pass locally
